+++ crates/iota-core/src/authority/authority_store.rs
@@ -1262,7 +1262,8 @@ impl AuthorityStore {
             .into_iter()
             .map(|(id, version)| ObjectKey(id, version));
 
-        // These versions no longer stopped being current in this epoch
+        // The revert makes the versions this transaction consumed current
+        // again.
         write_batch.delete_batch(
             &self.perpetual_tables.object_superseded_in_epoch,
             modified_object_keys.clone(),
