+++ crates/iota-core/src/unit_tests/abstract_account_test_utils.rs
@@ -5,15 +5,8 @@
 //!
 //! Publishes the same Move package the end-to-end abstract-account tests use,
 //! creates a shared account guarded by its ed25519 authenticator, and funds the
-//! account so it can pay for its own transactions.
-//!
-//! Account transactions are submitted straight to consensus, as
-//! `UserTransactionV1` or, with an attestation attached, `UserTransactionV2`.
-//! There are no certificates on this path: consensus ordering is what
-//! authorizes execution, and it is also where the transaction's shared inputs
-//! get their versions assigned. Assignment therefore happens at submission, so
-//! a test can rotate the account's key after building a transaction and have it
-//! execute against the rotated state.
+//! account so it can pay for its own transactions. The attestation and
+//! consensus-submission extensions live in `attestation_test_utils`.
 
 use std::{path::PathBuf, sync::Arc};
 
@@ -29,10 +22,8 @@ use iota_sdk_types::{
 };
 use iota_types::{
     IOTA_FRAMEWORK_PACKAGE_ID,
-    attestation::{Attestation, AttestationData, AttestedTransaction},
     crypto::{AccountKeyPair, get_account_key_pair},
     effects::{TransactionEffects, TransactionEffectsAPI},
-    messages_consensus::ConsensusTransaction,
     move_authenticator::MoveAuthenticator,
     move_package,
     object::Object,
@@ -43,15 +34,10 @@ use iota_types::{
     },
     utils::to_sender_signed_transaction,
 };
-use starfish_config::AuthorityIndex;
 
-use crate::{
-    authority::{
-        AuthorityState, authority_test_utils::send_and_confirm_transaction,
-        move_integration_tests::created_package_ref, test_authority_builder::TestAuthorityBuilder,
-    },
-    checkpoints::CheckpointServiceNoop,
-    consensus_handler::SequencedConsensusTransaction,
+use crate::authority::{
+    AuthorityState, authority_test_utils::send_and_confirm_transaction,
+    move_integration_tests::created_package_ref, test_authority_builder::TestAuthorityBuilder,
 };
 
 /// The abstract-account package shared with the end-to-end tests, relative to
@@ -71,12 +57,6 @@ const AA_KEYED_MODULE: &str = "abstract_account_keyed";
 const AA_AUTHENTICATE_ED25519: &str = "authenticate_ed25519";
 pub const AA_AUTHENTICATE_ED25519_VIA_SIGNING_DIGEST: &str =
     "authenticate_ed25519_via_signing_digest";
-
-/// Consensus accepts a validator attestation only from the block author, which
-/// `SequencedConsensusTransaction::new_test` fixes at index 0.
-fn attestor_index() -> AuthorityIndex {
-    AuthorityIndex::new_for_test(0)
-}
 
 pub struct AbstractAccountTestEnv {
     pub authority: Arc<AuthorityState>,
@@ -146,11 +126,11 @@ impl AbstractAccountTestEnv {
             .object_ref()
     }
 
-    fn rgp(&self) -> u64 {
+    pub(crate) fn rgp(&self) -> u64 {
         self.authority.reference_gas_price_for_testing().unwrap()
     }
 
-    fn budget(&self) -> u64 {
+    pub(crate) fn budget(&self) -> u64 {
         self.rgp() * TEST_ONLY_GAS_UNIT_FOR_PUBLISH * 10
     }
 
@@ -453,72 +433,4 @@ impl AbstractAccountTestEnv {
         );
     }
 
-    /// Runs the attestor's own dry-run, which is how a genuine attestation is
-    /// produced. Only succeeds while authentication still passes.
-    pub fn attest(&self, tx: &Transaction) -> Attestation {
-        let epoch_store = self.authority.epoch_store_for_testing();
-        let verified = epoch_store.verify_transaction(tx.clone()).unwrap();
-        let (payload, _) = self
-            .authority
-            .attest_transaction(&verified, &epoch_store)
-            .expect("attesting must succeed while authentication passes");
-        Attestation::Validator {
-            payload,
-            attestor_index: attestor_index(),
-        }
-    }
-
-    /// An attestation vouching for the transaction at the given versions, as a
-    /// dishonest attestor would produce. The computation estimate is the
-    /// largest one consensus accepts, so the verdict on the recorded versions
-    /// is never preempted by the attested-units cap on the re-run.
-    pub fn attest_with_versions(&self, object_versions: Vec<ObjectReference>) -> Attestation {
-        let computation_units = self.budget() / self.rgp();
-        Attestation::Validator {
-            payload: AttestationData::V1 {
-                computation_units,
-                object_versions,
-            },
-            attestor_index: attestor_index(),
-        }
-    }
-
-    /// Submits the transaction to consensus, with an attestation when one is
-    /// given, and executes what consensus schedules. Shared input versions are
-    /// assigned here, so the transaction runs against the account's state at
-    /// submission time rather than at the time it was built.
-    pub async fn submit(
-        &self,
-        tx: Transaction,
-        attestation: Option<Attestation>,
-    ) -> TransactionEffects {
-        let consensus_tx = match attestation {
-            Some(attestation) => ConsensusTransaction::new_user_transaction_v2(
-                AttestedTransaction::new(tx, attestation),
-            ),
-            None => ConsensusTransaction::new_user_transaction_v1(tx),
-        };
-        let epoch_store = self.authority.epoch_store_for_testing();
-        let scheduled = epoch_store
-            .process_consensus_transactions_for_tests(
-                vec![SequencedConsensusTransaction::new_test(consensus_tx)],
-                &Arc::new(CheckpointServiceNoop {}),
-                self.authority.get_object_cache_reader().as_ref(),
-                self.authority.get_transaction_cache_reader().as_ref(),
-                &self.authority.metrics,
-                true,
-                &self.authority,
-            )
-            .await
-            .unwrap();
-        let executable = scheduled
-            .into_iter()
-            .next()
-            .expect("consensus must schedule the submitted transaction");
-        let (effects, _) = self
-            .authority
-            .try_execute_immediately(&executable, None, &epoch_store)
-            .unwrap();
-        effects
-    }
 }
